# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/bill/purchases": {
      "get": {
        "operationId": "PurchasesController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contact_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "List purchase invoices with filtering and pagination",
        "tags": [
          "Bill"
        ]
      },
      "post": {
        "operationId": "PurchasesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Create a new purchase invoice",
        "tags": [
          "Bill"
        ]
      }
    },
    "/bill/purchases/stats": {
      "get": {
        "operationId": "PurchasesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Aggregate purchase stats — status counts + outstanding totals",
        "tags": [
          "Bill"
        ]
      }
    },
    "/bill/purchases/{id}": {
      "get": {
        "operationId": "PurchasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get purchase invoice details with items",
        "tags": [
          "Bill"
        ]
      },
      "patch": {
        "operationId": "PurchasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Update a purchase invoice (draft only)",
        "tags": [
          "Bill"
        ]
      }
    },
    "/bill/purchases/{id}/status": {
      "patch": {
        "operationId": "PurchasesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "sent"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Transition a purchase bill status (draft → sent / cancelled / …)",
        "tags": [
          "Bill"
        ]
      }
    },
...
    "/purchase-orders": {
      "post": {
        "operationId": "PurchaseOrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "contact_id",
                  "date",
                  "items"
                ],
                "properties": {
                  "contact_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "8b6b0c1e-2f5a-4b3c-9d8e-1a2b3c4d5e6f"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-06-01"
                  },
                  "delivery_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-06-10"
                  },
                  "shipping_address": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "place_of_supply": {
                    "type": "string",
                    "example": "29-Karnataka"
                  },
                  "is_igst": {
                    "type": "boolean",
                    "example": false
                  },
                  "notes": {
                    "type": "string",
                    "example": "Deliver to backside gate."
                  },
                  "terms": {
                    "type": "string",
                    "example": "Net 30 days."
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "description",
                        "quantity",
                        "rate"
                      ],
                      "properties": {
                        "product_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "description": {
                          "type": "string",
                          "example": "A4 Copier Paper - 80 GSM"
                        },
                        "hsn_code": {
                          "type": "string",
                          "example": "4802"
                        },
                        "quantity": {
                          "type": "number",
                          "example": 10
                        },
                        "unit": {
                          "type": "string",
                          "example": "reams"
                        },
                        "rate": {
                          "type": "number",
                          "example": 250
                        },
                        "discount_pct": {
                          "type": "number",
                          "example": 5
                        },
                        "cgst_rate": {
                          "type": "number",
                          "example": 9
                        },
                        "sgst_rate": {
                          "type": "number",
                          "example": 9
                        },
                        "igst_rate": {
                          "type": "number",
                          "example": 0
                        },
                        "cess_rate": {
                          "type": "number",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Create a new purchase order",
        "tags": [
          "Purchase Orders"
        ]
      },
      "get": {
        "operationId": "PurchaseOrdersController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "List purchase orders with filtering and pagination",
        "tags": [
          "Purchase Orders"
        ]
      }
    },
    "/purchase-orders/{id}": {
      "get": {
        "operationId": "PurchaseOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Get purchase order details with items",
        "tags": [
          "Purchase Orders"
        ]
      },
      "patch": {
        "operationId": "PurchaseOrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Partial purchase order fields to update."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Update a purchase order (draft/sent only)",
        "tags": [
          "Purchase Orders"
        ]
      },
      "delete": {
        "operationId": "PurchaseOrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Delete a draft purchase order",
        "tags": [
          "Purchase Orders"
        ]
      }
    },
    "/purchase-orders/{id}/status": {
      "patch": {
        "operationId": "PurchaseOrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "sent"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Update purchase order status",
        "tags": [
          "Purchase Orders"
        ]
      }
    },
    "/purchase-orders/{id}/convert-to-bill": {
      "post": {
        "operationId": "PurchaseOrdersController_convertToBill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "org-id": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Convert a purchase order to a purchase bill",
        "tags": [
          "Purchase Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kontafy API",
    "description": "REST API for Kontafy — cloud-native accounting for Indian businesses. All paths are versioned: prefix every request with `/v1`. Authenticated routes expect `Authorization: Bearer <access_token>` and most multi-tenant routes also expect `X-Org-Id: <organization id>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.kontafy.com/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:5002/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /v1/auth/login"
      },
      "org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Org-Id",
        "description": "Organization id for multi-tenant routes"
      }
    },
...
  }
}